+++ src/components/assignmentsPage/themes/AssignmentsDesktopPage.scss
@@ -119,6 +119,5 @@
 
 // overflow scroll filter classes sidebar
 .classrooms-filter {
-  height: calc(100vh - 5.834vw - 2vw - 5vw - 2.4vw - 2vw - 0.3vw);
   overflow-y: auto;
 }
