+++ src/components/play/PlayBrickRouting.tsx
@@ -86,6 +86,7 @@ import { SetFinishRedirectUrl, SetHeartOfMerciaUser, SetLoginRedirectUrl } from 
 import AdaptedBrickAssignedDialog from "./baseComponents/dialogs/AdaptedBrickAssignedDialog";
 import { AssignmentBrickStatus } from "model/assignment";
 import { getClassById } from "components/teach/service";
+import BottomAssignmentPopup from "./BottomAssignmentPopup";
 
 
 export enum PlayPage {
@@ -1102,24 +1103,6 @@ const BrickRouting: React.FC<BrickRoutingProps> = (props) => {
     return '';
   }
 
-  const renderCount = () => {
-    let assignmentsCount = assignClass?.assignments?.length;
-    if (assignmentsCount == 1) {
-      return (
-        <div>
-          <div className="class-name font-16"><span className="bold">{assignClass?.name}</span></div>
-          <div className="font-11">{assignmentsCount} Brick Assigned</div>
-        </div>
-      );
-    }
-    return (
-      <div>
-        <div className="class-name font-16"><span className="bold">{assignClass?.name} wefwefwef wefwef wef </span></div>
-        <div className="font-11">{assignmentsCount} Bricks Assigned</div>
-      </div>
-    );
-  }
-
   return (
     <React.Suspense fallback={<></>}>
       {isIPad13 || isTablet ? <TabletTheme /> : isMobile ? <MobileTheme /> : <DesktopTheme />}
@@ -1190,15 +1173,10 @@ const BrickRouting: React.FC<BrickRoutingProps> = (props) => {
       {renderCreditPopup()}
       {renderConvertBrills()}
       <AdaptedBrickAssignedDialog assignment={adaptedBrickAssignment} history={history} close={() => setAdaptedBrickAssignment(null)} />
-      {assignClass &&
-        <div className="bottom-bricks-popup-f53">
-          {renderCount()}
-          <div className="btn" onClick={() => {
-            setAssignClass(null);
-            props.history.push(props.history.location.href)
-          }}>Quit</div>
-          <div className="btn btn-green" onClick={() => props.history.push(map.TeachAssignedTab + '?classroomId=' + assignClass.id)}>Back to Class</div>
-        </div>}
+      {assignClass && <BottomAssignmentPopup click={() => {
+        setAssignClass(null);
+        props.history.push(props.history.location.href)
+      }} assignClass={assignClass} />}
     </React.Suspense>
   );
 };
